backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
azure-cognitiveservices-speech==1.35.0
supabase==2.3.4
python-multipart==0.0.6
pydantic==2.5.3
# 포먼트(공명) 분석용
librosa==0.10.1
numpy==1.26.3
scipy==1.12.0
praat-parselmouth==0.4.3
soundfile==0.12.1
